[refactor] #108 storeservice to interface
